# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Invoicing Database API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status?: string;
}

export interface Company {
  id: number;
  name: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Customer {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  zip?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  flatNumber?: string | null;
  /** @nullable */
  defaultRent?: number | null;
  createdAt: string;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  phone?: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  country?: string;
  notes?: string;
  flatNumber?: string;
  /** @minimum 0 */
  defaultRent?: number;
}

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  country?: string;
  notes?: string;
  flatNumber?: string;
  /** @minimum 0 */
  defaultRent?: number;
}

export type ProductUnit = typeof ProductUnit[keyof typeof ProductUnit];


export const ProductUnit = {
  each: 'each',
  hour: 'hour',
  day: 'day',
  month: 'month',
  flat: 'flat',
} as const;

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  unitPrice: number;
  unit: ProductUnit;
  createdAt: string;
}

export type ProductInputUnit = typeof ProductInputUnit[keyof typeof ProductInputUnit];


export const ProductInputUnit = {
  each: 'each',
  hour: 'hour',
  day: 'day',
  month: 'month',
  flat: 'flat',
} as const;

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  unitPrice: number;
  unit: ProductInputUnit;
}

export type ProductUpdateUnit = typeof ProductUpdateUnit[keyof typeof ProductUpdateUnit];


export const ProductUpdateUnit = {
  each: 'each',
  hour: 'hour',
  day: 'day',
  month: 'month',
  flat: 'flat',
} as const;

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  unitPrice?: number;
  unit?: ProductUpdateUnit;
}

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  draft: 'draft',
  sent: 'sent',
  paid: 'paid',
  overdue: 'overdue',
} as const;

export interface Invoice {
  id: number;
  invoiceNumber: string;
  customerId: number;
  status: InvoiceStatus;
  issueDate: string;
  dueDate: string;
  /** @nullable */
  notes?: string | null;
  subtotal: number;
  taxRate: number;
  taxAmount: number;
  total: number;
  createdAt: string;
  updatedAt: string;
}

export type InvoiceWithCustomerStatus = typeof InvoiceWithCustomerStatus[keyof typeof InvoiceWithCustomerStatus];


export const InvoiceWithCustomerStatus = {
  draft: 'draft',
  sent: 'sent',
  paid: 'paid',
  overdue: 'overdue',
} as const;

export interface InvoiceWithCustomer {
  id: number;
  invoiceNumber: string;
  customerId: number;
  customerName: string;
  status: InvoiceWithCustomerStatus;
  issueDate: string;
  dueDate: string;
  /** @nullable */
  notes?: string | null;
  subtotal: number;
  taxRate: number;
  taxAmount: number;
  total: number;
  createdAt: string;
  updatedAt: string;
}

export type InvoiceDetailStatus = typeof InvoiceDetailStatus[keyof typeof InvoiceDetailStatus];


export const InvoiceDetailStatus = {
  draft: 'draft',
  sent: 'sent',
  paid: 'paid',
  overdue: 'overdue',
} as const;

export interface LineItem {
  id: number;
  invoiceId: number;
  /** @nullable */
  productId?: number | null;
  description: string;
  quantity: number;
  unitPrice: number;
  amount: number;
}

export interface InvoiceDetail {
  id: number;
  invoiceNumber: string;
  customerId: number;
  customerName: string;
  /** @nullable */
  customerEmail: string | null;
  status: InvoiceDetailStatus;
  issueDate: string;
  dueDate: string;
  /** @nullable */
  notes?: string | null;
  subtotal: number;
  taxRate: number;
  taxAmount: number;
  total: number;
  lineItems: LineItem[];
  createdAt: string;
  updatedAt: string;
}

export type InvoiceInputStatus = typeof InvoiceInputStatus[keyof typeof InvoiceInputStatus];


export const InvoiceInputStatus = {
  draft: 'draft',
  sent: 'sent',
  paid: 'paid',
  overdue: 'overdue',
} as const;

export interface InvoiceInput {
  customerId: number;
  issueDate: string;
  dueDate: string;
  notes?: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  taxRate?: number;
  status?: InvoiceInputStatus;
}

export interface InvoiceUpdate {
  /** @minLength 1 */
  invoicePrefix?: string;
  /** @minLength 1 */
  invoiceNumber?: string;
  customerId?: number;
  issueDate?: string;
  dueDate?: string;
  notes?: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  taxRate?: number;
}

export type InvoiceStatusUpdateStatus = typeof InvoiceStatusUpdateStatus[keyof typeof InvoiceStatusUpdateStatus];


export const InvoiceStatusUpdateStatus = {
  draft: 'draft',
  sent: 'sent',
  paid: 'paid',
  overdue: 'overdue',
} as const;

export interface InvoiceStatusUpdate {
  status: InvoiceStatusUpdateStatus;
}

export interface InvoiceCustomField {
  label: string;
  text: string;
}

export interface InvoiceSettings {
  invoicePrefix: string;
  invoiceTitle: string;
  issuerName: string;
  issuerAddress: string;
  footerText: string;
  /** @nullable */
  logoUrl: string | null;
  customFields: InvoiceCustomField[];
  updatedAt: string;
}

export interface InvoiceSettingsUpdate {
  /** @minLength 1 */
  invoicePrefix?: string;
  invoiceTitle?: string;
  issuerName?: string;
  issuerAddress?: string;
  footerText?: string;
  /** @nullable */
  logoUrl?: string | null;
  customFields?: InvoiceCustomField[];
}

export interface InvoiceSummary {
  totalInvoiced: number;
  totalPaid: number;
  totalOutstanding: number;
  totalOverdue: number;
  invoiceCount: number;
  customerCount: number;
}

export interface LineItemInput {
  productId?: number;
  /** @minLength 1 */
  description: string;
  /** @minimum 0 */
  quantity: number;
  /** @minimum 0 */
  unitPrice: number;
}

export interface LineItemUpdate {
  productId?: number;
  /** @minLength 1 */
  description?: string;
  /** @minimum 0 */
  quantity?: number;
  /** @minimum 0 */
  unitPrice?: number;
}

export type ListCustomersParams = {
/**
 * Filter by name or email
 */
search?: string;
};

export type ListProductsParams = {
search?: string;
};

export type ListInvoicesParams = {
status?: ListInvoicesStatus;
customerId?: number;
search?: string;
};

export type ListInvoicesStatus = typeof ListInvoicesStatus[keyof typeof ListInvoicesStatus];


export const ListInvoicesStatus = {
  draft: 'draft',
  sent: 'sent',
  paid: 'paid',
  overdue: 'overdue',
} as const;