# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Invoicing Database API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Filter by name or email
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/invoices:
    get:
      operationId: getCustomerInvoices
      tags: [customers]
      summary: Get invoices for a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceWithCustomer"

  # ── Products ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Customer ──────────────────────────────────────────────────────────────
    Customer:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zip:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        flatNumber:
          type: ["string", "null"]
        defaultRent:
          type: ["number", "null"]
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string
        notes:
          type: string
        flatNumber:
          type: string
        defaultRent:
          type: number
          minimum: 0
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string
        notes:
          type: string
        flatNumber:
          type: string
        defaultRent:
          type: number
          minimum: 0

    # ── Product ───────────────────────────────────────────────────────────────
...
    InvoiceWithCustomer:
      type: object
      required: [id, invoiceNumber, customerId, customerName, status, issueDate, dueDate, subtotal, taxRate, taxAmount, total, createdAt, updatedAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        issueDate:
          type: string
        dueDate:
          type: string
        notes:
          type: ["string", "null"]
        subtotal:
          type: number
        taxRate:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
...